Moving common errors to main package
