+++ CONTRIBUTING.md
@@ -7,7 +7,7 @@ Thank you for your interest in helping develop Hord. The time, skills, and persp
 Bugs, Design Proposals, Feature Requests, and Questions are all welcome and can be submitted by creating a [Github Issue](https://github.com/madflojo/hord/issues/new/choose) using one of the templates provided. Please provide as much detail as you can.
 
 Code contributions are welcome as well! In an effort to keep this project tidy, please:
-- Use `dep` to install and lock dependencies
+- Use `go mod` to install and lock dependencies
 - Use `gofmt` to format code and tests
 - Run `go vet -v ./...` to check for any inadvertent suspicious code
 - Write and run unit tests when they make sense using `go test`
