+++ drivers/cassandra/cassandra_test.go
@@ -2,6 +2,7 @@ package cassandra
 
 import (
 	"fmt"
+	"github.com/madflojo/hord"
 	"testing"
 	"time"
 )
@@ -9,32 +10,32 @@ import (
 func TestErrNoDial(t *testing.T) {
 	var db Database
 	err := db.Setup()
-	if err != ErrNoDial {
+	if err != hord.ErrNoDial {
 		t.Errorf("Expected no dialing error but got - %s", err)
 	}
 
 	err = db.HealthCheck()
-	if err != ErrNoDial {
+	if err != hord.ErrNoDial {
 		t.Errorf("Expected no dialing error but got - %s", err)
 	}
 
 	err = db.Set("key", []byte("test"))
-	if err != ErrNoDial {
+	if err != hord.ErrNoDial {
 		t.Errorf("Expected no dialing error but got - %s", err)
 	}
 
 	_, err = db.Get("key")
-	if err != ErrNoDial {
+	if err != hord.ErrNoDial {
 		t.Errorf("Expected no dialing error but got - %s", err)
 	}
 
 	err = db.Delete("key")
-	if err != ErrNoDial {
+	if err != hord.ErrNoDial {
 		t.Errorf("Expected no dialing error but got - %s", err)
 	}
 
 	_, err = db.Keys()
-	if err != ErrNoDial {
+	if err != hord.ErrNoDial {
 		t.Errorf("Expected no dialing error but got - %s", err)
 	}
 }
